requirements.txt
certifi==2018.8.24
Click==7.0
dj-database-url==0.5.0
Django==2.1.4
django-heroku==0.3.1
Flask==1.0.2
Flask-Session==0.3.1
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.14.0
opencv-python==3.4.0.12
pandas==0.22.0
pdfkit==0.6.1
psycopg2==2.7.6.1
python-dateutil==2.6.1
pytz==2018.3
six==1.11.0
SQLAlchemy==1.2.16
virtualenv==16.0.0
Werkzeug==0.14.1
whitenoise==4.1.2
xlrd==1.1.0
